Allow single-homed <-> multi-homed transitions
To change the cluster from single homed to multi homed or vice versa one
must target the master node first, and pass the --force option. All
other nodes then will work as long as they are reachable by the master.

Note that this will also prevent a node to be set to single-homed if the
master is multi-homed, which wasn't disallowed before, and warn if a
single-homed <-> multi-homed transition happens.

Also note that it's still theoretically possible to flip a cluster
inadvertently by changing the master node this way, and then doing a
master failover before fixing the other nodes.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>